Reworked structural flow of game.
Originally planned to use a screenloader object to handle logic of each
screen within the same file. Issues popped up and I also realized this
won't scale very well if we have like 40 screens. Decided each screen
will be an object that inherits from a Screen class. Each screen will
have a load method that will initialize the current screen and an update
method that will be called inside the main update method when that
screen is active.